+++ src/promo/stories/articleDonationBanner.jsx
@@ -1,7 +1,10 @@
-import React from 'react'
-import ComponentOverview from '../../ComponentOverview'
 import ArticleDonationBanner, { ArticleDonationBanner as UnwrappedArticleDonationBanner } from '../ArticleDonationBanner'
+import ComponentOverview from '../../ComponentOverview'
+import React from 'react'
 import { Avatar, Person, Typography } from '../../index'
+import { action } from '@storybook/addon-actions'
+
+import koala from './koala.png'
 
 export default () => (
   <ComponentOverview heading='ArticleDonationBanner' component={UnwrappedArticleDonationBanner}>
@@ -9,7 +12,7 @@ export default () => (
       <h2>Example</h2>
     </div>
 
-    <ArticleDonationBanner href='http://donate.theconverastion.com' donateText='Donate now'>
+    <ArticleDonationBanner onClick={action('clicked')} donateText='Donate now'>
       <Typography variant='h6' gutterBottom>Before you go...</Typography>
       <Typography variant='body1' paragraph>
         The Conversation serves society by making knowledge accessible to everyone, not just a select few. Our only agenda is a better informed public. If you care about what we do please make a donation now and help secure our future.
@@ -18,7 +21,7 @@ export default () => (
         <Typography gutterBottom>
           <Person name='Misha Ketchell' caption='Editor' />
         </Typography>
-        <Avatar size={48} />
+        <Avatar size={48} src={koala} />
       </div>
     </ArticleDonationBanner>
   </ComponentOverview>