+++ src/promo/stories/donationBanner.jsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import ComponentOverview from '../../ComponentOverview'
 import DonationBanner, { DonationBanner as UnwrappedDonationBanner } from '../DonationBanner'
+import { action } from '@storybook/addon-actions'
 
 export default () => (
   <ComponentOverview heading='DonationBanner' component={UnwrappedDonationBanner}>
@@ -8,7 +9,7 @@ export default () => (
       <h2>Example</h2>
     </div>
 
-    <DonationBanner href='http://donate.theconverastion.com'>
+    <DonationBanner onClick={action('clicked')}>
       The Conversation provides clean information in a world infected with spin. Please donate.
     </DonationBanner>
   </ComponentOverview>